add install CUDA host driver
Container using images with driver version >= 10.2 need similar driver on the
host. CUDA is usually backward-compatible. So it is OK to install the latest
version on the host and the containers use equally new or older version. The
reverse doesn't work: Newer driver version in the containers is not runnable
if host driver is older.