Make Global build optional for covid19_nextstrain.sh
In order to build Global build of the same data, rerun the exact command you ran for
Denmark with additional `-g` flag. It is crucial to use the same input metadata,
sequences file, as well as ouput directory so the program can reuse the results
of 'Denmark' build.

If you want to build with Global from start, use `-g` flag from beginning.